+++ src/features/editor/extensions/code-block-container/code-block/index.ts
@@ -7,8 +7,6 @@ import { PrismPlugin } from "./prism-plugin";
 
 export interface CodeBlockOptions {
   languageClassPrefix: string;
-  exitOnTripleEnter: boolean;
-  exitOnArrowDown: boolean;
   defaultLanguage: string;
   HTMLAttributes: Record<string, any>;
 }
@@ -19,8 +17,6 @@ export const CodeBlock = Node.create<CodeBlockOptions>({
   addOptions() {
     return {
       languageClassPrefix: "language-",
-      exitOnTripleEnter: true,
-      exitOnArrowDown: true,
       defaultLanguage: "plaintext",
       HTMLAttributes: {},
     };
@@ -85,7 +81,7 @@ export const CodeBlock = Node.create<CodeBlockOptions>({
       dom.className = "code-block-wrapper-for-langname"; // 言語名表示のポジションのため
       dom.setAttribute(
         "data-language",
-        node.attrs.language || this.options.defaultLanguage,
+        node.attrs.language || this.options.defaultLanguage
       );
       const pre = document.createElement("pre");
 
@@ -133,8 +129,8 @@ export const CodeBlock = Node.create<CodeBlockOptions>({
               $from.after(-1),
               this.editor.state.schema.nodes.paragraph.create(
                 null,
-                text ? [this.editor.state.schema.text(text)] : [],
-              ),
+                text ? [this.editor.state.schema.text(text)] : []
+              )
             );
 
             return true;
@@ -146,10 +142,6 @@ export const CodeBlock = Node.create<CodeBlockOptions>({
 
       // exit node on triple enter
       Enter: ({ editor }) => {
-        if (!this.options.exitOnTripleEnter) {
-          return false;
-        }
-
         const { state } = editor;
         const { selection } = state;
         const { $from, empty } = selection;
@@ -178,106 +170,14 @@ export const CodeBlock = Node.create<CodeBlockOptions>({
           })
           .run();
       },
-
-      // exit node on arrow down
-      ArrowDown: ({ editor }) => {
-        if (!this.options.exitOnArrowDown) {
-          return false;
-        }
-
-        const { state } = editor;
-        const { selection, doc } = state;
-        const { $from, empty } = selection;
-
-        if (!empty || $from.parent.type !== this.type) {
-          return false;
-        }
-
-        const isAtEnd = $from.parentOffset === $from.parent.nodeSize - 2;
-
-        if (!isAtEnd) {
-          return false;
-        }
-
-        const after = $from.after();
-
-        if (after === undefined) {
-          return false;
-        }
-
-        const nodeAfter = doc.nodeAt(after);
-
-        if (nodeAfter) {
-          return editor.commands.command(({ tr }) => {
-            tr.setSelection(Selection.near(doc.resolve(after)));
-            return true;
-          });
-        }
-
-        return editor.commands.exitCode();
-      },
     };
   },
 
-  // TODO: 動作するようにする
   addProseMirrorPlugins() {
     return [
       PrismPlugin({
         name: this.name,
         defaultLanguage: this.options.defaultLanguage,
-      }),
-      new Plugin({
-        key: new PluginKey("codeBlockVSCodeHandler"),
-        props: {
-          handlePaste: (view, event) => {
-            if (!event.clipboardData) {
-              return false;
-            }
-
-            // don’t create a new code block within code blocks
-            if (this.editor.isActive(this.type.name)) {
-              return false;
-            }
-
-            const text = event.clipboardData.getData("text/plain");
-            const vscode = event.clipboardData.getData("vscode-editor-data");
-            const vscodeData = vscode ? JSON.parse(vscode) : undefined;
-            const language = vscodeData?.mode;
-
-            if (!text || !language) {
-              return false;
-            }
-
-            const { tr, schema } = view.state;
-
-            // prepare a text node
-            // strip carriage return chars from text pasted as code
-            // see: https://github.com/ProseMirror/prosemirror-view/commit/a50a6bcceb4ce52ac8fcc6162488d8875613aacd
-            const textNode = schema.text(text.replace(/\r\n?/g, "\n"));
-
-            // create a code block with the text node
-            // replace selection with the code block
-            tr.replaceSelectionWith(this.type.create({ language }, textNode));
-
-            if (tr.selection.$from.parent.type !== this.type) {
-              // put cursor inside the newly created code block
-              tr.setSelection(
-                TextSelection.near(
-                  tr.doc.resolve(Math.max(0, tr.selection.from - 2)),
-                ),
-              );
-            }
-
-            // store meta information
-            // this is useful for other plugins that depends on the paste event
-            // like the paste rule plugin
-            tr.setMeta("paste", true);
-
-            view.dispatch(tr);
-
-            return true;
-          },
-        },
       }),
     ];
   },
